+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## main
 
+BREAKING CHANGES:
+
+- The deprecated `contact_id` field has been removed from the `dnsimple_lets_encrypt_certificate` resource. (dnsimple/terraform-provider-dnsimple#93)
+- The `id` and `domain_id` fields on the `dnsimple_lets_encrypt_certificate` are now of type `int64` instead of `string`. (dnsimple/terraform-provider-dnsimple#93)
+- The `domain_id` field on the `dnsimple_lets_encrypt_certificate` is now required. (dnsimple/terraform-provider-dnsimple#93)
+
 ## 0.16.3
 
 BUG FIXES: